+++ Software/Install Adobe Reader DC.ps1
@@ -64,31 +64,37 @@ Write-Host -ForegroundColor Green "...ready"
 Write-Output ""
 
 # Check, if a new version is available
-[version]$Version = Get-Content -Path "$PSScriptRoot\$Product\Version.txt"
-[version]$Adobe = (Get-ItemProperty HKLM:\Software\Wow6432Node\Microsoft\Windows\CurrentVersion\Uninstall\* | Where-Object {$_.DisplayName -like "*Adobe Acrobat Reader*"}).DisplayVersion
-IF ($Adobe -lt $Version) {
-# Adobe Reader DC Update
-Write-Host -ForegroundColor Yellow "Installing Adobe Reader DC Update"
-DS_WriteLog "I" "Installing Adobe Reader DC Update" $LogFile
-try {
-	$mspArgs = "/P `"$PSScriptRoot\$Product\Adobe_DC_MUI_Update.msp`" /quiet /qn"
-	Start-Process -FilePath msiexec.exe -ArgumentList $mspArgs -Wait	
-	} catch {
-DS_WriteLog "E" "Error while installing Adobe Reader DC Update (error: $($Error[0]))" $LogFile       
-}
-DS_WriteLog "-" "" $LogFile
+IF (Test-Path -Path "$PSScriptRoot\$Product\Version.txt") {
+	[version]$Version = Get-Content -Path "$PSScriptRoot\$Product\Version.txt"
+	[version]$Adobe = (Get-ItemProperty HKLM:\Software\Wow6432Node\Microsoft\Windows\CurrentVersion\Uninstall\* | Where-Object {$_.DisplayName -like "*Adobe Acrobat Reader*"}).DisplayVersion
+	IF ($Adobe -lt $Version) {
+	# Adobe Reader DC Update
+	Write-Host -ForegroundColor Yellow "Installing Adobe Reader DC Update"
+	DS_WriteLog "I" "Installing Adobe Reader DC Update" $LogFile
+	try {
+		$mspArgs = "/P `"$PSScriptRoot\$Product\Adobe_DC_MUI_Update.msp`" /quiet /qn"
+		Start-Process -FilePath msiexec.exe -ArgumentList $mspArgs -Wait	
+		} catch {
+	DS_WriteLog "E" "Error while installing Adobe Reader DC Update (error: $($Error[0]))" $LogFile       
+	}
+	DS_WriteLog "-" "" $LogFile
 
-# Disale update service and scheduled task
-Start-Sleep 5
-Stop-Service AdobeARMservice
-Set-Service AdobeARMservice -StartupType Disabled
-Disable-ScheduledTask -TaskName "Adobe Acrobat Update Task" | Out-Null
-Write-Host -ForegroundColor Green "...ready"
-Write-Output ""
-}
+	# Disale update service and scheduled task
+	Start-Sleep 5
+	Stop-Service AdobeARMservice
+	Set-Service AdobeARMservice -StartupType Disabled
+	Disable-ScheduledTask -TaskName "Adobe Acrobat Update Task" | Out-Null
+	Write-Host -ForegroundColor Green "...ready"
+	Write-Output ""
+	}
 
-# Stop, if no new version is available
+	# Stop, if no new version is available
+	Else {
+	Write-Host "No Update available for $Product"
+	Write-Output ""
+	}
+}
 Else {
-Write-Host "No Update available for $Product"
+Write-Host -ForegroundColor Red "Version file not found for $Product"
 Write-Output ""
 }
