+++ Software/Install NotepadPlusPlus.ps1
@@ -48,27 +48,33 @@ DS_WriteLog "-" "" $LogFile
 #========================================================================================================================================
 
 # Check, if a new version is available
-[version]$Version = Get-Content -Path "$PSScriptRoot\$Product\Version.txt"
-[version]$Notepad = (Get-ItemProperty HKLM:\Software\Microsoft\Windows\CurrentVersion\Uninstall\* | Where-Object {$_.DisplayName -like "*Notepad++*"}).DisplayVersion
-# $Notepad = $Notepad.TrimEnd('.0')
-IF ($Notepad -lt $Version) {
+IF (Test-Path -Path "$PSScriptRoot\$Product\Version.txt") {
+	[version]$Version = Get-Content -Path "$PSScriptRoot\$Product\Version.txt"
+	[version]$Notepad = (Get-ItemProperty HKLM:\Software\Microsoft\Windows\CurrentVersion\Uninstall\* | Where-Object {$_.DisplayName -like "*Notepad++*"}).DisplayVersion
+	# $Notepad = $Notepad.TrimEnd('.0')
+	IF ($Notepad -lt $Version) {
 
-# Installation Notepad++
-Write-Host -ForegroundColor Yellow "Installing $Product"
-DS_WriteLog "I" "Installing $Product" $LogFile
-try	{
-	Start-Process "$PSScriptRoot\$Product\NotePadPlusPlus_x64.exe" –ArgumentList '/S /NoUpdater' –NoNewWindow -Wait
-	Remove-Item -Path "C:\Program Files\Notepad++\updater" -Recurse -EA SilentlyContinue
-	} catch {
-DS_WriteLog "E" "Error installing $Product (error: $($Error[0]))" $LogFile       
-}
-DS_WriteLog "-" "" $LogFile
-Write-Host -ForegroundColor Green " ...ready!" 
-Write-Output ""
-}
+	# Installation Notepad++
+	Write-Host -ForegroundColor Yellow "Installing $Product"
+	DS_WriteLog "I" "Installing $Product" $LogFile
+	try	{
+		Start-Process "$PSScriptRoot\$Product\NotePadPlusPlus_x64.exe" –ArgumentList '/S /NoUpdater' –NoNewWindow -Wait
+		Remove-Item -Path "C:\Program Files\Notepad++\updater" -Recurse -EA SilentlyContinue
+		} catch {
+	DS_WriteLog "E" "Error installing $Product (error: $($Error[0]))" $LogFile       
+	}
+	DS_WriteLog "-" "" $LogFile
+	Write-Host -ForegroundColor Green " ...ready!" 
+	Write-Output ""
+	}
 
-# Stop, if no new version is available
+	# Stop, if no new version is available
+	Else {
+	Write-Host "No Update available for $Product"
+	Write-Output ""
+	}
+}
 Else {
-Write-Host "No Update available for $Product"
+Write-Host -ForegroundColor Red "Version file not found for $Product"
 Write-Output ""
 }
